crossdev: drop {binutils,gcc}-config forced upgrade
These packages have been in stable for a while now, so there shouldn't be
much of a need to force a sanity upgrade.  Especially since the common
toolchain packages take care of forcing recent ones.

Any stragglers will simply be made fun of if they attempt to use crossdev
on an outdated system.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>