crossdev: add overlay options for people to control sources
The current auto-search aspect of crossdev can be confusing to people, and
creating the overlay by hand is a pain for others, so add some command line
options for people to explicitly control the overlays that crossdev uses.

Now you can tell crossdev where to search in general, where to find exact
packages, and where to write its result -- all from the command line.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>