crossdev: go back to setting up multilib, but smarter
The recent commit 48942e883df62 didn't operate exactly the way I wanted
and allowed the host profile's idea of multilib to bleed in for most
packages (but not the cross-compiler ones where multilib.eclass fixed
things up).  So leverage multilib.eclass to avoid duplicating the env
func locally and to setup the ABI env as we want.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>