crossdev: make sure to always process extra pkgs
For targets which do not go through all stages (like arm-non-eabi), we
end up always exiting early before we get a chance to process the extra
stuff (like gdb).  So allow the extra args to always be processed.

That's the only functional change here.  The rest is finally using proper
if statements and indenting relevant blocks of code.  It's grown too much
at this point to get away with the previous `foo && bar` logic.

Reported-by: David Hendricks <dhendrix@google.com>
Signed-off-by: Mike Frysinger <vapier@gentoo.org>