cross-pkg-config: remember PKG_CONFIG from build time
This should avoid recursive cross-pkg-config calls when someone sets
CBUILD = CHOST in the cross-emerge environment, which might be useful
for building a system image with SYSROOT != /.

For sys-devel/crossdev we will call "tc-export PKG_CONFIG" to set a
reasonable default for REAL_PKG_CONFIG. This will generally be
${CHOST}-pkg-config, where CBUILD = CHOST and SYSROOT = ROOT = /.

In the case we are cross-compiling crossdev itself, we want
${CHOST}-pkg-config for REAL_PKG_CONFIG in SYSROOT/ROOT in case it is later
used as a root image on appropriate hardware.

We also allow the user to override REAL_PKG_CONFIG via the environment
in case some unforseen use case is discovered.

Bug: https://bugs.gentoo.org/955822
Signed-off-by: Mike Gilbert <floppym@gentoo.org>