crossdev: rework nvptx support
Split nvptx support into two:
* nvptx-none for `nvtpx-none-gcc` and friends which retain the ability
  to build standalone code for nvptx.

* accel-nvptx-none for the acclerator bits for GCC.

To get up and running, two commands will now be needed:
$ crossdev nvptx-none
$ crossdev accel-nvptx-none

This approach means that upgrades and rebuilds of components will work
fine, as opposed to before where newlib or nvptx-tools could not be rebuilt
once the process had completed. The other options evaluated are detailed
in the bug.

Bug: https://bugs.gentoo.org/965844
Signed-off-by: Sam James <sam@gentoo.org>